+++ src/lib/index.ts
@@ -2,6 +2,7 @@ import cookie from 'cookie'
 import { SuperFetch } from 'sveltekit-superfetch'
 import type { ProductDTO } from '@medusajs/types'
 import type { Cookies, RequestEvent } from '@sveltejs/kit'
+import { dev } from '$app/environment'
 
 export interface ProductRetrievalOptions {
    limit?: number
@@ -77,10 +78,11 @@ export interface ClientOptions {
    timeout?: number
    headers?: {}
    persistentCart?: boolean
-   cache?: boolean
-   debug?: boolean
+   disableCache?: boolean
+   ttl?: number
    logger?: Logger
    logFormat?: 'text' | 'json' | 'majel'
+   logLevel?: 'verbose' | 'limited' | 'silent'
    excludedPaths?: string[]
    limitedPaths?: string[]
 }
@@ -90,6 +92,9 @@ export interface QueryOptions {
    path: string
    method?: string
    body?: object
+   key?: string
+   ttl?: number
+   logLevel?: 'verbose' | 'limited' | 'silent'
 }
 
 export interface Logger {
@@ -103,10 +108,11 @@ export class MedusaClient {
    private retry: number = 0
    private headers: any
    private persistentCart: boolean = false
-   private cache: boolean = false
-   private debug: boolean = false
+   private disableCache: boolean = false
+   private ttl: number = 1000
    private logger: Logger = console
-   private logFormat: 'text' | 'json' | 'majel' = 'text'
+   private logFormat: 'text' | 'json' | 'majel' = 'json'
+   private logLevel: 'verbose' | 'limited' | 'silent' = (dev)? 'limited' : 'silent'
    private excludedPaths: string[] = ['/store/auth']
    private limitedPaths: string[] = []
    private superFetch: SuperFetch
@@ -114,15 +120,16 @@ export class MedusaClient {
    constructor(url: string, options: ClientOptions = {}) {
       this.url = url
       if (options) {
-         let { timeout, retry, headers, persistentCart, cache, debug, logger, logFormat, excludedPaths, limitedPaths } = options
+         let { timeout, retry, headers, persistentCart, disableCache, ttl, logger, logFormat, logLevel, excludedPaths, limitedPaths } = options
          if (timeout) this.timeout = timeout
          if (retry) this.retry = retry
          if (headers) this.headers = headers
          if (persistentCart) this.persistentCart = persistentCart
-         if (cache) this.cache = cache
-         if (debug) this.debug = debug
+         if (disableCache) this.disableCache = disableCache
+         if (ttl) this.ttl = ttl
          if (logger) this.logger = logger
          if (logFormat) this.logFormat = logFormat
+         if (logLevel) this.logLevel = logLevel
          if (excludedPaths) { 
             for (const excluded of excludedPaths) {
                this.excludedPaths.push(excluded) 
@@ -137,17 +144,17 @@ export class MedusaClient {
       this.superFetch = new SuperFetch({
          retry: this.retry,
          timeout: this.timeout,
-         debug: this.debug,
-         cache: this.cache,
+         ttl: this.ttl,
          logger: this.logger,
          logFormat: this.logFormat,
+         logLevel: this.logLevel,
          excludedPaths: this.excludedPaths,
          limitedPaths: this.limitedPaths
       })
    }
 
    async query(options: QueryOptions): Promise<Response|null> {
-      const { locals, path, method = 'GET', body = {} } = options
+      const { locals, path, method = 'GET', body = {}, ...rest } = options
       let headers: any = {}
       if (this.headers) {
         for (const [key, value] of Object.entries(this.headers)) {
@@ -164,8 +171,9 @@ export class MedusaClient {
          url: `${this.url}${path}`,
          method,
          headers,
-         body: (Object.keys(body).length != 0) ? JSON.stringify(body) : null
-      }).catch((e) => {
+         body: (Object.keys(body).length != 0) ? JSON.stringify(body) : null,
+         ...rest
+      }).catch((e: Error) => {
          console.log(e)
          return null
       })
@@ -240,7 +248,8 @@ export class MedusaClient {
          locals, 
          path: '/store/auth', 
          method: 'POST', 
-         body: { email, password }
+         body: { email, password },
+         logLevel: 'silent'
       })
       if (!response || !response.ok) return false
       return await this.parseAuthCookie(response?.headers?.getSetCookie(), locals, cookies).catch(() => false)
@@ -267,7 +276,8 @@ export class MedusaClient {
          locals, 
          path: '/store/customers', 
          method: 'POST', 
-         body: user
+         body: user,
+         logLevel: 'silent'
       }).then((res:any) => {
          if (res.ok) { 
             return this.login(locals, cookies, email, password).then(() => true).catch(() => false) 
@@ -278,44 +288,51 @@ export class MedusaClient {
    async getSearchResults(q:string) {
       // returns an array of hits, if any
       if (!q) { return Array() }
+      let key: string|undefined = (this.disableCache)? undefined: `${q}_search`
       return await this.query({
          path: '/store/products/search',
          method: 'POST',
-         body: { q }
+         body: { q },
+         key
       }).then((res:any) => res.json()).then((data:any) => data.hits).catch(() => null)
    }
 
    async getProducts(options:ProductRetrievalOptions = {}): Promise<ProductDTO[]|null> {
       // returns an array of product objects
+      let key: string|undefined = (this.disableCache)? undefined: 'all_products'
       const queryString = this.buildQuery('/store/products', options)
-      return await this.query({ path: queryString })
+      return await this.query({ path: queryString, key })
       .then((res:any) => res.json()).then((data:any) => data.products).catch(() => null)
    }
 
    async getCollections(options:CollectionRetrievalOptions = {}) {
       // returns an array of collection objects on success
+      let key: string|undefined = (this.disableCache)? undefined: 'all_collections'
       const queryString = this.buildQuery('/store/collections', options)
-      return await this.query({ path: queryString })
+      return await this.query({ path: queryString, key })
       .then((res:any) => res.json()).then((data:any) => data.collections).catch(() => null)
    }
 
    async getCollection(handle:string) {
       // returns a collection object on success
-      return await this.query({ path: `/store/collections?handle[]=${handle}` })
+      let key: string|undefined = (this.disableCache)? undefined: `${handle}_collection`
+      return await this.query({ path: `/store/collections?handle[]=${handle}`, key })
       .then((res:any) => res.json()).then((data:any) => data.collections[0]).catch(() => null)
    }
 
    async getCollectionProducts(id:string, options:ProductRetrievalOptions = {}): Promise<ProductDTO[]|null> {
       // returns an array of product objects on success
+      let key: string|undefined = (this.disableCache)? undefined: `${id}_products`
       let base = `/store/products?collection_id[]=${id}`
       const queryString = this.buildQuery(base, options)
-      return await this.query({ path: queryString })
+      return await this.query({ path: queryString, key })
       .then((res:any) => res.json()).then((data:any) => data.products).catch(() => null)
    }
 
    async getProduct(handle:string): Promise<ProductDTO|null> {
       // returns a product object on success
-      let product = await this.query({ path: `/store/products?handle=${handle}` })
+      let key: string|undefined = (this.disableCache)? undefined: `${handle}_product`
+      let product = await this.query({ path: `/store/products?handle=${handle}`, key })
       .then((res:any) => res.json()).then((data:any) => data.products[0]).catch(() => null)
       if (!product) { return null }
       for (let option of product.options) {
@@ -328,7 +345,8 @@ export class MedusaClient {
       // returns an array of review objects on success
       // options - page = 1, limit = 10, sort = 'created_at', order = 'desc', search = null
       // TODO: handle options
-      return await this.query({ path: `/store/products/${productId}/reviews` })
+      let key: string|undefined = (this.disableCache)? undefined: `${productId}_reviews`
+      return await this.query({ path: `/store/products/${productId}/reviews`, key })
       .then((res:any) => res.json()).then((data:any) => data.product_reviews).catch(() => null)
    }